refactor(webgl): use a single canvas for bg & fg
