+++ src/renderer/render/webgl/renderer.ts
@@ -1,10 +1,19 @@
 import CreateWebGLBuffer from './grid-buffer'
-import { cell } from '../../workspace'
-import TextFG from './text-fg'
-import TextBG from './text-bg'
 import { cursor as cursorState } from '../../cursor'
-import { CursorShape } from '../../../common/types'
 import { getActiveGridId } from '../../windows/window-manager'
+import { getColorAtlas } from '../highlight-attributes'
+import generateFontAtlas from '../font-texture-atlas'
+import { cell } from '../../workspace'
+import { CursorShape } from '../../../common/types'
+import * as twgl from 'twgl.js'
+// @ts-ignore
+import fgVertShader from './shaders/text-fg-vert.glsl'
+// @ts-ignore
+import fgFragShader from './shaders/text-fg-frag.glsl'
+// @ts-ignore
+import bgVertShader from './shaders/text-bg-vert.glsl'
+// @ts-ignore
+import bgFragShader from './shaders/text-bg-frag.glsl'
 
 export interface WebGLView {
   resize: (rows: number, cols: number) => void
@@ -31,58 +40,448 @@ export interface WebGLView {
   resetAtlasBounds: () => void
 }
 
-const nutella = () => {
-  let fontAtlasCache: HTMLCanvasElement
+const createRenderer = () => {
+  const canvas = document.createElement('canvas') as HTMLCanvasElement
+  const gl = canvas.getContext('webgl2', { alpha: true, preserveDrawingBuffer: true })
+  if (!gl) throw new Error('couldn\'t create webgl context . . . hmm')
+  twgl.addExtensionsToContext(gl)
+  gl.pixelStorei(gl.UNPACK_PREMULTIPLY_ALPHA_WEBGL, true)
 
-  let textFGRenderer = TextFG()
-  let textBGRenderer = TextBG()
+  const fgProgramInfo = twgl.createProgramInfo(gl, [fgVertShader, fgFragShader])
+  const bgProgramInfo = twgl.createProgramInfo(gl, [bgVertShader, bgFragShader])
 
-  const resizeCanvas = (width: number, height: number) => {
-    textBGRenderer.resize(width, height)
-    textFGRenderer.resize(width, height)
+  const viewport = { x: 0, y: 0, width: 0, height: 0 }
+  let shouldShowCursor = false
+  let cursorShape = CursorShape.block
+ 
+  gl.enable(gl.SCISSOR_TEST)
+  gl.enable(gl.BLEND)
+  gl.blendFunc(gl.SRC_ALPHA, gl.ONE_MINUS_SRC_ALPHA)
+
+  const colorAtlas = getColorAtlas()
+  const colorAtlasTex = twgl.createTexture(gl, { src: colorAtlas })
+
+  const uniforms = {
+    canvasResolution: [0, 0],
+    fontAtlasResolution: [0, 0],
+    colorAtlasResolution: [colorAtlas.width, colorAtlas.height],
+    fontAtlasTextureId: 0,
+    colorAtlasTextureId: colorAtlasTex,
+    cellSize: [cell.width, cell.height],
+    cellPadding: [0, cell.padding],
+
+    hlidType: 0,
+
+    shouldShowCursor: true,
+    cursorPosition: [0, 0],
+    cursorShape: 0,
+    cursorColor: [0, 0, 0, 1],
+    bgCursorColor: [1, 1, 1, 1],
   }
 
-  const showCursor = (enable: boolean) => {
-    textBGRenderer.showCursor(enable)
-    textFGRenderer.showCursor(enable)
+  // wait for roboto-mono to be loaded before we generate the initial font atlas
+  ;(document as any).fonts.ready.then(() => {
+    const fontAtlas = generateFontAtlas()
+    const fontAtlasWidth = Math.floor(fontAtlas.width / window.devicePixelRatio)
+    const fontAtlasHeight = Math.floor(
+      fontAtlas.height / window.devicePixelRatio
+    )
+
+    Object.assign(uniforms, { 
+      fontAtlasTextureId: twgl.createTexture(gl, { src: fontAtlas }),
+      fontAtlasResolution: [fontAtlasWidth, fontAtlasHeight ]
+    })
+  })
+
+  const renderBuffer = gl.createBuffer()
+  const quadBuffer = gl.createBuffer()
+  const bgQuadBuffer = gl.createBuffer()
+  if (!renderBuffer || !quadBuffer || !bgQuadBuffer) throw new Error("umm . . .") // TODO(smolck)
+
+  const stride = 7 * Float32Array.BYTES_PER_ELEMENT
+  const bufferInfo: twgl.BufferInfo = {
+    numElements: 0, // TODO(smolck)
+    attribs: {
+      cellPosition: {
+        buffer: renderBuffer,
+        size: 2,
+        type: gl.FLOAT,
+        offset: 0,
+        stride,
+        divisor: 1,
+      },
+      hlid: {
+        buffer: renderBuffer,
+        size: 1,
+        type: gl.FLOAT,
+        offset: 2 * Float32Array.BYTES_PER_ELEMENT,
+        stride,
+        divisor: 1,
+      },
+      isSecondHalfOfDoubleWidthCell: {
+        buffer: renderBuffer,
+        size: 1,
+        type: gl.FLOAT,
+        offset: 4 * Float32Array.BYTES_PER_ELEMENT,
+        stride,
+        divisor: 1,
+      },
+      atlasBounds: {
+        buffer: renderBuffer,
+        size: 2,
+        type: gl.FLOAT,
+        offset: 5 * Float32Array.BYTES_PER_ELEMENT,
+        stride,
+        divisor: 1,
+      },
+      quadVertex: {
+        buffer: quadBuffer,
+        type: gl.FLOAT,
+        size: 2
+      },
+
+      bgQuadVertex: {
+        buffer: bgQuadBuffer,
+        type: gl.FLOAT,
+        size: 2
+      },
+      isCursorTri: {
+        buffer: bgQuadBuffer,
+        type: gl.FLOAT,
+        size: 1,
+        offset: Float32Array.BYTES_PER_ELEMENT * 2 * 12,
+      },
+    }
   }
 
-  const updateCursorShape = (shape: CursorShape) => {
-    textBGRenderer.updateCursorShape(shape)
-    // TODO(smolck): If cursor size changes need to update cells . . .
-    textBGRenderer.updateCellSize(false, cursorState.size)
+  gl.bindBuffer(gl.ARRAY_BUFFER, quadBuffer)
+  gl.bufferData(gl.ARRAY_BUFFER, new Float32Array([
+      0,
+      0,
+      cell.width,
+      cell.height,
+      0,
+      cell.height,
+      cell.width,
+      0,
+      cell.width,
+      cell.height,
+      0,
+      0,
+  ]), gl.STATIC_DRAW)
 
-    textFGRenderer.updateCursorShape(shape)
+  const updateCellSize = (initial = false, cursorSize = 20) => {
+    const w = cell.width
+    const h = cell.height
+    const smallerW = w * (cursorSize / 100.0)
+    const percentH = h * (cursorSize / 100.0)
+
+    const next = {
+      boxes: new Float32Array([
+        0,
+        0,
+        smallerW,
+        h,
+        0,
+        h,
+
+        smallerW,
+        0,
+        smallerW,
+        h,
+        0,
+        0,
+
+        smallerW,
+        0,
+        w,
+        h,
+        smallerW,
+        h,
+
+        w,
+        0,
+        w,
+        h,
+        smallerW,
+        0,
+
+        // TODO(smolck): Better way of doing this? Also, note that the 1's
+        // specify which triangles of the above to color in for the cursor, and the zeroes
+        // which triangles not to color in, *if* the cursor is a line shape. If
+        // it isn't a line shape (atm a block shape), these are ignored.
+        ...Array(6).fill(1),
+        ...Array(6).fill(0),
+      ]),
+      // TODO(smolck): Don't draw double the tris for underliens too, maybe use
+      // a separate buffer for quadVertex somehow or something?
+      lines: new Float32Array([
+        /* Previous values (for future ref):
+         * 0, cell.height - 1,
+         * cell.width, cell.height,
+         * 0, cell.height,
+         *
+         * cell.width, cell.height - 1,
+         * cell.width, cell.height,
+         * 0, cell.height - 1, */
+        0,
+        h - 1,
+        smallerW,
+        percentH,
+        0,
+        percentH,
+
+        smallerW,
+        h - 1,
+        smallerW,
+        percentH,
+        0,
+        h - 1,
+
+        smallerW,
+        h - 1,
+        w,
+        percentH,
+        smallerW,
+        percentH,
+
+        w,
+        h - 1,
+        w,
+        percentH,
+        smallerW,
+        h - 1,
+
+        ...Array(12).fill(0),
+      ]),
+    }
+
+    uniforms.cellSize = [cell.width, cell.height]
+    if (!initial) Object.assign(quads, next)
+
+    gl.bindBuffer(gl.ARRAY_BUFFER, quadBuffer)
+    gl.bufferData(gl.ARRAY_BUFFER, new Float32Array([
+        0,
+        0,
+        cell.width,
+        cell.height,
+        0,
+        cell.height,
+        cell.width,
+        0,
+        cell.width,
+        cell.height,
+        0,
+        0,
+    ]), gl.STATIC_DRAW)
+
+    Object.assign(uniforms, {
+      cellSize: [cell.width, cell.height],
+      cellPadding: [0, cell.padding]
+    })
+    return next
   }
 
-  const updateCursorColor = (r: number, g: number, b: number) => {
-    textBGRenderer.updateCursorColor([r, g, b])
-    textFGRenderer.updateCursorColor([1.0 - r, 1.0 - g, 1.0 - b])
+  const quads = updateCellSize(true)
+
+  const fgVertexArrayInfo = twgl.createVertexArrayInfo(gl, fgProgramInfo, bufferInfo)
+  const bgVertexArrayInfo = twgl.createVertexArrayInfo(gl, bgProgramInfo, bufferInfo)
+
+  const resize = (width: number, height: number) => {
+    const w = Math.round(width * window.devicePixelRatio)
+    const h = Math.round(height * window.devicePixelRatio)
+
+    if (canvas.width !== w || canvas.height !== h) {
+      canvas.width = w
+      canvas.height = h
+      canvas.style.width = `${width}px`
+      canvas.style.height = `${height}px`
+    }
   }
 
-  const updateCursorPosition = (row: number, col: number) => {
-    textBGRenderer.updateCursorPosition(row, col)
-    textFGRenderer.updateCursorPosition(row, col)
+  const readjustViewportMaybe = (
+    x: number,
+    y: number,
+    width: number,
+    height: number
+  ) => {
+    const bottom = (y + height) * window.devicePixelRatio
+    const yy = Math.round(canvas.height - bottom)
+    const xx = Math.round(x * window.devicePixelRatio)
+    const ww = Math.round(width * window.devicePixelRatio)
+    const hh = Math.round(height * window.devicePixelRatio)
+
+    const same =
+      viewport.width === ww &&
+      viewport.height === hh &&
+      viewport.x === xx &&
+      viewport.y === yy
+
+    if (same) return
+
+    Object.assign(viewport, { x: xx, y: yy, width: ww, height: hh })
+    gl.viewport(viewport.x, viewport.y, viewport.width, viewport.height)
+    gl.scissor(viewport.x, viewport.y, viewport.width, viewport.height)
+    Object.assign(uniforms, { canvasResolution: [width, height] })
+  }
+
+  const render = (
+    buffer: Float32Array,
+    x: number,
+    y: number,
+    width: number,
+    height: number
+  ) => {
+    readjustViewportMaybe(x, y, width, height)
+    gl.bindBuffer(gl.ARRAY_BUFFER, renderBuffer)
+    gl.bufferData(gl.ARRAY_BUFFER, buffer, gl.STATIC_DRAW)
+
+    const renderFg = () => {
+      gl.useProgram(fgProgramInfo.program)
+      gl.bindVertexArray(fgVertexArrayInfo.vertexArrayObject!!)
+      twgl.setUniforms(fgProgramInfo, uniforms)
+      twgl.drawBufferInfo(gl, bufferInfo, gl.TRIANGLES, 6, 0, buffer.length / 7)
+    }
+
+    const renderBg = () => {
+      gl.useProgram(bgProgramInfo.program)
+      gl.bindVertexArray(bgVertexArrayInfo.vertexArrayObject!!)
+      // uniforms.cursorColor = uniforms.cursorColor.slice(0, 2).reverse()
+
+      // background
+      if (shouldShowCursor && cursorShape == 2) uniforms.shouldShowCursor = false
+      uniforms.hlidType = 0
+      twgl.setUniforms(bgProgramInfo, uniforms)
+      gl.bindBuffer(gl.ARRAY_BUFFER, bgQuadBuffer)
+      gl.bufferData(gl.ARRAY_BUFFER, quads.boxes, gl.STATIC_DRAW)
+      twgl.drawBufferInfo(gl, bufferInfo, gl.TRIANGLES, 12, 0, buffer.length / 7)
+
+      // underlines
+      uniforms.shouldShowCursor = shouldShowCursor
+      gl.bindBuffer(gl.ARRAY_BUFFER, bgQuadBuffer)
+      gl.bufferData(gl.ARRAY_BUFFER, quads.lines, gl.STATIC_DRAW)
+
+      // Just want to ignore the cursor logic in the vertex shader for underlines,
+      // so set shouldShowCursor to false, then back to it's previous value after
+      // the draw call.
+      if (shouldShowCursor && cursorShape != CursorShape.underline)
+        uniforms.shouldShowCursor = false
+      uniforms.hlidType = 2
+      twgl.setUniforms(bgProgramInfo, uniforms)
+      twgl.drawBufferInfo(gl, bufferInfo, gl.TRIANGLES, 12, 0, buffer.length / 7)
+
+      uniforms.shouldShowCursor = shouldShowCursor
+    }
+
+    renderBg()
+    renderFg()
   }
 
   const updateFontAtlas = (fontAtlas: HTMLCanvasElement) => {
-    textFGRenderer.updateFontAtlas(fontAtlas)
+    const width = Math.floor(fontAtlas.width / window.devicePixelRatio)
+    const height = Math.floor(fontAtlas.height / window.devicePixelRatio)
+    Object.assign(uniforms, {
+      fontAtlasResolution: [width, height],
+      fontAtlasTextureId: twgl.createTexture(gl, { src: fontAtlas })
+    })
+  }
+
+  const showCursor = (enable: boolean) => 
+  ((shouldShowCursor = enable), (uniforms.shouldShowCursor = enable))
+
+  const updateCursorColor = (color: [number, number, number]) => {
+    const [r, g, b] = color
+    uniforms.cursorColor = [1.0 - r, 1.0 - g, 1.0 - b, 1]
+    uniforms.bgCursorColor = [r, g, b, 1]
+  }
+
+  const updateCursorShape = (shape: CursorShape) => {
+    cursorShape = shape
+    uniforms.cursorShape = shape
+  }
+
+  const updateCursorPosition = (row: number, col: number) => {
+    uniforms.cursorPosition = [col, row]
+  }
+
+  const updateColorAtlas = (colorAtlas: HTMLCanvasElement) => {
+    uniforms.colorAtlasTextureId = twgl.createTexture(gl, { src: colorAtlas })
+    uniforms.colorAtlasResolution = [colorAtlas.width, colorAtlas.height]
+  }
+
+  const clear = (x: number, y: number, width: number, height: number) => {
+    readjustViewportMaybe(x, y, width, height)
+    gl.clear(gl.COLOR_BUFFER_BIT)
+  }
+
+  const clearAll = () => {
+    readjustViewportMaybe(
+      0,
+      0,
+      canvas.clientWidth,
+      canvas.clientHeight
+    )
+    gl.clear(gl.COLOR_BUFFER_BIT)
+  }
+
+  return {
+    clear,
+    clearAll,
+    render,
+    resize,
+    updateFontAtlas,
+    updateColorAtlas,
+    updateCellSize,
+    updateCursorPosition,
+    updateCursorShape,
+    updateCursorColor,
+    showCursor,
+    canvasElement: canvas,
+  }
+}
+
+const nutella = () => {
+  let fontAtlasCache: HTMLCanvasElement
+
+  let renderer = createRenderer()
+
+  const resizeCanvas = (width: number, height: number) => {
+    renderer.resize(width, height)
+  }
+
+  const showCursor = (enable: boolean) => {
+    renderer.showCursor(enable)
+  }
+
+  const updateCursorShape = (shape: CursorShape) => {
+    renderer.updateCursorShape(shape)
+    // TODO(smolck): If cursor size changes need to update cells . . .
+    renderer.updateCellSize(false, cursorState.size)
+  }
+
+  const updateCursorColor = (r: number, g: number, b: number) => {
+    renderer.updateCursorColor([r, g, b])
+  }
+
+  const updateCursorPosition = (row: number, col: number) => {
+    renderer.updateCursorPosition(row, col)
+  }
+
+  const updateFontAtlas = (fontAtlas: HTMLCanvasElement) => {
+    renderer.updateFontAtlas(fontAtlas)
     fontAtlasCache = fontAtlas
   }
 
   const updateCellSize = () => {
-    textBGRenderer.updateCellSize()
-    textFGRenderer.updateCellSize()
+    renderer.updateCellSize()
   }
 
   const updateColorAtlas = (colorAtlas: HTMLCanvasElement) => {
-    textBGRenderer.updateColorAtlas(colorAtlas)
-    textFGRenderer.updateColorAtlas(colorAtlas)
+    renderer.updateColorAtlas(colorAtlas)
   }
 
   const clearAll = () => {
-    textBGRenderer.clearAll()
-    textFGRenderer.clearAll()
+    renderer.clearAll()
   }
 
   const createView = (initialGridId: number): WebGLView => {
@@ -126,8 +525,7 @@ const nutella = () => {
 
       const doHacks = gridId !== getActiveGridId() && cursorState.visible
       if (doHacks) showCursor(false)
-      textBGRenderer.render(buffer, x, y, width, height)
-      textFGRenderer.render(buffer, x, y, width, height)
+      renderer.render(buffer, x, y, width, height)
       if (doHacks) showCursor(true)
     }
 
@@ -137,15 +535,13 @@ const nutella = () => {
 
       const doHacks = gridId !== getActiveGridId() && cursorState.visible
       if (doHacks) showCursor(false)
-      textBGRenderer.render(buffer, x, y, width, height)
-      textFGRenderer.render(buffer, x, y, width, height)
+      renderer.render(buffer, x, y, width, height)
       if (doHacks) showCursor(true)
     }
 
     const clear = () => {
       const { x, y, width, height } = viewport
-      textBGRenderer.clear(x, y, width, height)
-      textFGRenderer.clear(x, y, width, height)
+      renderer.clear(x, y, width, height)
     }
 
     const clearGridBuffer = () => gridBuffer.clear()
@@ -154,16 +550,14 @@ const nutella = () => {
       gridBuffer.moveRegionUp(lines, top, bottom)
       const buffer = gridBuffer.getBuffer()
       const { x, y, width, height } = viewport
-      textBGRenderer.render(buffer, x, y, width, height)
-      textFGRenderer.render(buffer, x, y, width, height)
+      renderer.render(buffer, x, y, width, height)
     }
 
     const moveRegionDown = (lines: number, top: number, bottom: number) => {
       gridBuffer.moveRegionDown(lines, top, bottom)
       const buffer = gridBuffer.getBuffer()
       const { x, y, width, height } = viewport
-      textBGRenderer.render(buffer, x, y, width, height)
-      textFGRenderer.render(buffer, x, y, width, height)
+      renderer.render(buffer, x, y, width, height)
     }
 
     return {
@@ -197,15 +591,10 @@ const nutella = () => {
     updateCursorPosition,
     updateCursorColor,
     showCursor,
-    foregroundElement: textFGRenderer.canvasElement,
-    backgroundElement: textBGRenderer.canvasElement,
-    reInit: ({ fg, bg }: { fg: boolean; bg: boolean }) => {
-      if (fg) {
-        textFGRenderer = TextFG()
-        updateFontAtlas(fontAtlasCache)
-      }
-      if (bg) textBGRenderer = TextBG()
-
+    canvasElement: renderer.canvasElement,
+    reInit: () => {
+      renderer = createRenderer()
+      updateFontAtlas(fontAtlasCache)
       showCursor(cursorState.visible)
       updateCursorPosition(cursorState.row, cursorState.col)
       updateCursorShape(cursorState.shape)
