+++ src/renderer/render/webgl/shaders/text-bg-vert.glsl
@@ -1,5 +1,5 @@
 #version 300 es
-in vec2 quadVertex;
+in vec2 bgQuadVertex;
 in vec2 cellPosition;
 in float isCursorTri;
 in float hlid;
@@ -8,7 +8,7 @@ uniform vec2 cursorPosition;
 uniform vec2 canvasResolution;
 uniform vec2 colorAtlasResolution;
 uniform vec2 cellSize;
-uniform vec4 cursorColor;
+uniform vec4 bgCursorColor;
 uniform bool shouldShowCursor;
 uniform int cursorShape;
 uniform float hlidType;
@@ -21,7 +21,7 @@ void main() {
   bool colorInSecondHalfOfDoubleWidthCell = isSecondHalfOfDoubleWidthCell == 1.0 && cursorPosition == prevCellPos && cursorShape == 0;
   bool isCursorCell = (colorInSecondHalfOfDoubleWidthCell || cursorPosition == cellPosition) && shouldShowCursor;
 
-  vec2 vertexPosition = (cellPosition * cellSize) + quadVertex;
+  vec2 vertexPosition = (cellPosition * cellSize) + bgQuadVertex;
   vec2 posFloat = vertexPosition / canvasResolution;
   gl_Position = vec4(posFloat.x * 2.0 - 1.0, posFloat.y * -2.0 + 1.0, 0, 1);
 
@@ -34,7 +34,7 @@ void main() {
   // check if we're in the cursor cell and if we should color in this triangle;
   // otherwise, just check if we're in the cursor cell.
   if (cursorShape == 1 ? isCursorCell && isCursorTri == 1.0 : isCursorCell) {
-    o_color = cursorColor;
+    o_color = bgCursorColor;
   } else {
     vec4 textureColor = texture(colorAtlasTextureId, colorPosition);
     o_color = textureColor;
